+++ libs/ratanpy/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 ### Added
-- **Fast Acquisition 1-3 GHz:** added borrowed noise interpolator, large nan-gaps channel filter
+- **Fast Acquisition 1-3 GHz:** borrowed noise interpolator
+- **Fast Acquisition 1-3 GHz:** large nan-gaps channel filter
+- **Fast Acquisition 1-3 GHz:** time-axis and frequency-axis downsamplers
 
 ## [2.0.7] - 2026-07-10
 ### Changed